+++ sdk/test/configuration/yaml_test.cc
@@ -176,7 +176,9 @@ attribute_limits:
   ASSERT_NE(config, nullptr);
   ASSERT_EQ(config->file_format, "1.0");
   ASSERT_NE(config->attribute_limits, nullptr);
-  ASSERT_EQ(config->attribute_limits->attribute_value_length_limit, 4096);
+  ASSERT_EQ(config->attribute_limits->attribute_value_length_limit,
+            opentelemetry::sdk::configuration::AttributeLimitsConfiguration::
+                kDefaultAttributeValueLengthLimit);
   ASSERT_EQ(config->attribute_limits->attribute_count_limit, 128);
 }
 
@@ -608,7 +610,9 @@ tracer_provider:
   auto ratio_sampler =
       static_cast<opentelemetry::sdk::configuration::TraceIdRatioBasedSamplerConfiguration *>(
           sampler);
-  ASSERT_EQ(ratio_sampler->ratio, 0.0);
+  ASSERT_EQ(
+      ratio_sampler->ratio,
+      opentelemetry::sdk::configuration::TraceIdRatioBasedSamplerConfiguration::kDefaultRatio);
 }
 
 TEST(Yaml, illegal_double)
@@ -653,7 +657,9 @@ tracer_provider:
   auto ratio_sampler =
       static_cast<opentelemetry::sdk::configuration::TraceIdRatioBasedSamplerConfiguration *>(
           sampler);
-  ASSERT_EQ(ratio_sampler->ratio, 0.0);
+  ASSERT_EQ(
+      ratio_sampler->ratio,
+      opentelemetry::sdk::configuration::TraceIdRatioBasedSamplerConfiguration::kDefaultRatio);
 }
 
 TEST(Yaml, empty_double_substitution)
@@ -680,7 +686,9 @@ tracer_provider:
   auto ratio_sampler =
       static_cast<opentelemetry::sdk::configuration::TraceIdRatioBasedSamplerConfiguration *>(
           sampler);
-  ASSERT_EQ(ratio_sampler->ratio, 0.0);
+  ASSERT_EQ(
+      ratio_sampler->ratio,
+      opentelemetry::sdk::configuration::TraceIdRatioBasedSamplerConfiguration::kDefaultRatio);
 }
 
 TEST(Yaml, with_double_substitution)